apps: move helpers closer to their usage
